python/socketio-fastapi/requirements.txt
fastapi
uvicorn
python-socketio
aiohttp
